feat: ARM-88 finishing touches and documentation for calendar input
