+++ module/src/components/calendarDisplay/index.ts
@@ -1,0 +1,2 @@
+export * from './calendarDisplay.component';
+export * from './calendarDisplay.utils';
