--- module/src/components/calendarInteraction/index.ts
@@ -1,1 +1,0 @@
-export * from './calendarInteraction.component';
