+++ module/src/components/calendarView/index.ts
@@ -1,2 +1,1 @@
 export * from './calendarView.component';
-export * from './calendarView.utils';
