+++ module/src/index.ts
@@ -9,8 +9,8 @@ export { Calendar, Form };
 export * from './components/autoCompleteInput';
 export * from './components/autoCompleteInputMulti';
 export * from './components/button';
+export * from './components/calendarDisplay';
 export * from './components/calendarInput';
-export * from './components/calendarInteraction';
 export * from './components/calendarView';
 export * from './components/characterLimit';
 export * from './components/checkboxInput';
