+++ playground/src/views/examples/calendar.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { CalendarInput, Form } from "@rocketmakers/armstrong-edge"
 
 export const CalendarExample: React.FC = () => {
-  const { formState, formProp } = Form.use<{ myDate: string }>()
+  const { formProp } = Form.use<{ myDate: string }>()
 
-  return <CalendarInput bind={formProp("myDate").bind()} formatString="dd/MM/yyyy" />
+  return <CalendarInput bind={formProp("myDate").bind()} formatString="dd/MM/yyyy" displayMode="calendar" />
 }
